add schemaInterface and remove ConstraintFinderTrait.php & ConstraintFinderInterface.php
